Added Batman event (3%) Reorganised packages for events
